requirements.txt
apache-airflow==3.0.0
pandas==2.2.2
psycopg2-binary==2.9.9
google-cloud-bigquery
